+++ hermes/__init__.py
@@ -1,0 +1,4 @@
+"""Top-level package for Hermes."""
+
+__all__ = ["core", "ui", "services", "data"]
+__version__ = "0.1.0"
