+++ hermes/core/__init__.py
@@ -1,1 +1,3 @@
 """Core functionality of Hermes."""
+
+__all__ = ["registro_ideias"]
