+++ hermes/services/__init__.py
@@ -1,0 +1,3 @@
+"""External service integrations for Hermes."""
+
+__all__ = ["llm_interface"]
