+++ hermes/ui/__init__.py
@@ -1,0 +1,3 @@
+"""User interfaces for Hermes."""
+
+__all__ = ["cli", "gui"]
