+++ pyproject.toml
@@ -1,0 +1,18 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "hermes"
+version = "0.1.0"
+description = "Hermes - Registro de ideias com LLM"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "PyQt5>=5.15.9",
+    "requests",
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["hermes*"]
