+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-# Ensure the parent directory is on the path so the Hermes package can be imported.
+# Ensure the parent directory is on the path so the hermes package can be imported.
 sys.path.append(
-    os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+    os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 )
