+++ tests/test_llm_interface.py
@@ -12,11 +12,11 @@ if 'requests' not in sys.modules:
     requests_stub.post = lambda *a, **k: None
     sys.modules['requests'] = requests_stub
 
-from Hermes import llm_interface
+from hermes.services import llm_interface
 
 
 class TestGerarResposta(unittest.TestCase):
-    @patch('Hermes.llm_interface.requests.post')
+    @patch('hermes.services.llm_interface.requests.post')
     def test_resposta_sucesso(self, mock_post):
         mock_response = Mock()
         mock_response.json.return_value = {"response": "ok"}
@@ -28,7 +28,7 @@ class TestGerarResposta(unittest.TestCase):
         )
         self.assertEqual(result, "ok")
 
-    @patch('Hermes.llm_interface.requests.post')
+    @patch('hermes.services.llm_interface.requests.post')
     def test_falha_conexao(self, mock_post):
         mock_post.side_effect = llm_interface.requests.exceptions.ConnectionError("falha")
         result = llm_interface.gerar_resposta(
@@ -36,7 +36,7 @@ class TestGerarResposta(unittest.TestCase):
         )
         self.assertTrue(result.startswith("[FALHA]"))
 
-    @patch('Hermes.llm_interface.requests.post')
+    @patch('hermes.services.llm_interface.requests.post')
     def test_resposta_inesperada(self, mock_post):
         mock_response = Mock()
         mock_response.json.return_value = {"foo": "bar"}
